+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     * One of them has two wires
     * Had to rewire everything to use parasitic mode
     * `RESIN_HOST_CONFIG_dtoverlay=w1-gpio,pullup=1`
+* Temperature
+    * Report `t=85000` as error (sensor reset, error, ...)    
 
 ## v0.0.1 (2018-06-20)